requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.9.2
psycopg2-binary>=2.8
python-dateutil>=2.8.1
python-geohash>=0.8.5
requests
git+https://github.com/m-wrzr/populartimes
django-crontab>=0.7.1
geopy
beautifulsoup4>=4.9.3

